Main: hide ShadowTextureManager
